Change the HighlightGroup component to a public component.
