+++ packages/client/src/components/highlight/index.css
@@ -1,0 +1,47 @@
+.as-highlight-card {
+    position: relative;
+    border-radius: 0.75rem;
+    border: 1px solid var(--border);
+    background-color: var(--background);
+    box-shadow: inset 0 0 0 1px rgba(148, 163, 184, 0.08);
+    overflow: hidden;
+    transition:
+        border-color 0.25s ease,
+        box-shadow 0.25s ease,
+        background-color 0.25s ease;
+    --cursor-x: 50%;
+    --cursor-y: 50%;
+    --cursor-opacity: 0;
+}
+
+.as-highlight-card::before {
+    content: '';
+    position: absolute;
+    inset: 0;
+    padding: 1px;
+    border-radius: inherit;
+    background: radial-gradient(
+        200px circle at var(--cursor-x) var(--cursor-y),
+        rgba(0, 0, 0, 0.8),
+        rgba(0, 0, 0, 0.4) 45%,
+        transparent 82%
+    );
+    opacity: var(--cursor-opacity, 0);
+    pointer-events: none;
+    transition: opacity 0.25s ease;
+    box-sizing: border-box;
+    z-index: 0;
+    -webkit-mask:
+        linear-gradient(#fff 0 0) content-box,
+        linear-gradient(#fff 0 0);
+    -webkit-mask-composite: xor;
+    mask:
+        linear-gradient(#fff 0 0) content-box,
+        linear-gradient(#fff 0 0);
+    mask-composite: exclude;
+}
+
+.as-highlight-card > * {
+    position: relative;
+    z-index: 1;
+}
